+++ client/src/components/index.tsx
@@ -4,9 +4,7 @@ import Brand from "./Brand";
 import Loader from "./Loader";
 import Hotel from "./Hotel";
 import Hero from "./Hero";
-import AddBrand from "./AddBrand";
 import UpdateBrand from "./UpdateBrand";
-import AddHotel from "./AddHotel";
 import UpdateHotel from "./updateHotel";
 
 export {
@@ -16,8 +14,6 @@ export {
   Brand,
   Hotel,
   Hero,
-  AddBrand,
   UpdateBrand,
-  AddHotel,
   UpdateHotel,
 };
