+++ client/src/pages/AddBrand.tsx
@@ -1,0 +1,63 @@
+import { useMutation } from "@apollo/client";
+import React, { useState } from "react";
+import { CREATE_HOTEL_BRAND } from "../mutations/createBrand";
+import { GET_HOTEL_BRANDS } from "../queries/brands";
+
+const AddBrand = () => {
+  const [name, setName] = useState("");
+  const user = JSON.parse(localStorage.getItem("profile") || "{}");
+  const checkUser = Object.keys(user).length;
+  const [createBrand] = useMutation(CREATE_HOTEL_BRAND);
+  const [errors, setErrors] = useState("");
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+
+    if (name === "") {
+      return alert("Please fill in the field");
+    }
+
+    if (checkUser === 0) {
+      return setErrors("You must be logged in");
+    }
+
+    createBrand({
+      variables: {
+        input: {
+          name: name,
+        },
+      },
+      refetchQueries: [{ query: GET_HOTEL_BRANDS }],
+    });
+
+    setName("");
+  };
+
+  return (
+    <>
+      <div className="container">
+        <p className="text-center text-danger fs-3 mb-1 font-weight-semibold">
+          {errors}
+        </p>
+        <form onSubmit={handleSubmit}>
+          <div className="mb-3">
+            <label className="form-label">Brand Name</label>
+            <input
+              type="text"
+              id="name"
+              className="form-control"
+              value={name}
+              onChange={(e) => setName(e.target.value)}
+            />
+          </div>
+
+          <button className="btn btn-secondary" type="submit">
+            Submit
+          </button>
+        </form>
+      </div>
+    </>
+  );
+};
+
+export default AddBrand;
