+++ client/src/pages/index.tsx
@@ -3,5 +3,7 @@ import SinglePage from "./SinglePage";
 import Login from "./Login";
 import Register from "./Register";
 import SingleHotel from "./SingleHotel";
+import AddHotel from "./AddHotel";
+import AddBrand from "./AddBrand";
 
-export { Home, SinglePage, Login, Register, SingleHotel };
+export { Home, SinglePage, Login, Register, SingleHotel, AddHotel, AddBrand };
